refactor: 기존 TokenService 대체
