+++ src/main/java/nextstep/auth/ui/controller/AuthController.java
@@ -4,8 +4,6 @@ import nextstep.auth.application.dto.AuthRequest;
 import nextstep.auth.application.dto.AuthResponse;
 import nextstep.auth.application.dto.GithubAuthRequest;
 import nextstep.auth.application.service.AuthService;
-import nextstep.member.application.dto.TokenRequest;
-import nextstep.member.application.dto.TokenResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
