--- src/main/java/nextstep/member/application/dto/TokenResponse.java
@@ -1,16 +1,0 @@
-package nextstep.member.application.dto;
-
-public class TokenResponse {
-    private String accessToken;
-
-    public TokenResponse() {
-    }
-
-    public TokenResponse(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-}
