--- src/main/java/nextstep/member/ui/controller/TokenController.java
@@ -1,33 +1,0 @@
-package nextstep.member.ui.controller;
-
-import nextstep.member.application.service.TokenService;
-import nextstep.member.application.dto.GithubAccessTokenRequest;
-import nextstep.member.application.dto.TokenRequest;
-import nextstep.member.application.dto.TokenResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TokenController {
-    private TokenService tokenService;
-
-    public TokenController(TokenService tokenService) {
-        this.tokenService = tokenService;
-    }
-
-    @PostMapping("/login/token")
-    public ResponseEntity<TokenResponse> createToken(@RequestBody TokenRequest request) {
-        TokenResponse response = tokenService.createToken(request.getEmail(), request.getPassword());
-
-        return ResponseEntity.ok(response);
-    }
-
-    @PostMapping("/login/github")
-    public ResponseEntity<TokenResponse> createGithubToken(@RequestBody GithubAccessTokenRequest request) {
-        TokenResponse response = tokenService.createGithubToken(request.getCode());
-
-        return ResponseEntity.ok(response);
-    }
-}
