+++ src/test/java/nextstep/auth/application/GithubClientTest.java
@@ -1,6 +1,6 @@
 package nextstep.auth.application;
 
-import nextstep.member.application.dto.GithubProfileResponse;
+import nextstep.auth.application.dto.GithubProfileResponse;
 import nextstep.auth.utils.fixture.GithubAuthFixture;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
